move fields module to forms.fields to distinguish from model fields
