+++ application/forms/other_person_health_check/dob_auth.py
@@ -2,7 +2,7 @@ from datetime import date
 
 from django import forms
 
-from application.fields import CustomSplitDateFieldDOB
+from application.forms.fields import CustomSplitDateFieldDOB
 from application.forms.childminder import ChildminderForms
